logparser/requirements.txt
fastapi
uvicorn
pymongo
pydantic
pydantic-settings
python-dotenv
pika
motor
dnspython
httpx
PyYAML